pyproject.toml
[project]
name = "domowik"
version = "0.1.0"
description = ""
authors = [
    {name = "Igor Chistruga",email = "igor.chistruga@endava.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "dishka (>=1.5.3,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "typer (>=0.15.3,<0.16.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    "arq (>=0.26.3,<0.27.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
packages = [{include = "app"}]

[tool.ruff]
line-length=200

[tool.ruff.lint]
select = ["ALL"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.9"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
